change seed.js to generate ratings instead of usersLiked, and moved location of test/index.js to tasks/test.js, and abstracted helpers from tasks/test.js and tasks/seed.js to tasks/helpers.js
