+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "node ./src/app.js",
     "seed": "node ./tasks/seed.js",
-    "test": "node ./test/index.js",
+    "test": "node ./tasks/test.js",
     "format:check": "npx prettier --check .",
     "format:write": "npx prettier --write ."
   },
